requirements.txt
aiofiles==23.1.0
anyio==3.6.2
bcrypt==4.0.1
click==8.1.3
fastapi==0.95.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector-python==8.0.33
passlib==1.7.4
Pillow==9.5.0
protobuf==3.20.3
pydantic==1.10.7
PyJWT==2.6.0
pyotp==2.8.0
pypng==0.20220715.0
python-decouple==3.7
python-multipart==0.0.6
qrcode==7.4.2
sniffio==1.3.0
SQLAlchemy==2.0.12
starlette==0.26.1
typing_extensions==4.5.0
uvicorn==0.22.0
